fix cast

+++ internal/api/middleware/prometheus.go
@@ -16,6 +16,6 @@ func PrometheusMiddleware() gin.HandlerFunc {
 		status := ctx.Writer.Status()
 		metrics.HttpDuration.WithLabelValues(
 			path, method, strconv.Itoa(status),
-		).Observe(float64(time.Since(start)) / float64(time.Millisecond))
+		).Observe(float64(time.Since(start) / time.Millisecond))
 	}
 }
